move pd controller to separated pkg
