+++ pkg/backup/safe_point.go
@@ -60,7 +60,7 @@ func MakeSafePointID() string {
 // CheckGCSafePoint checks whether the ts is older than GC safepoint.
 // Note: It ignores errors other than exceed GC safepoint.
 func CheckGCSafePoint(ctx context.Context, pdClient pd.Client, ts uint64) error {
-	// TODO: use PDClient.GetGCSafePoint instead once PD client exports it.
+	// TODO: use pdClient.GetGCSafePoint instead once PD client exports it.
 	safePoint, err := getGCSafePoint(ctx, pdClient)
 	if err != nil {
 		log.Warn("fail to get GC safe point", zap.Error(err))
